backend/requirements.txt
Django==5.1.1
mysqlclient==2.2.4
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.1
virtualenv==20.26.5
django-cors-headers==4.4.0
djangorestframework==3.15.2